+++ src/AppNavigator.js
@@ -1,7 +1,7 @@
 import { createStackNavigator, createAppContainer } from 'react-navigation';
 import WelcomePage from './Components/WelcomePage/WelcomePage';
 import HomePage from './Components/HomePage/HomePage';
-import Routine from './Components/Routine/Routine'
+import Routine from './Components/RoutineCreator/RoutineCreator'
 
 const rootStack = createStackNavigator({
   welcomePage:  { screen: WelcomePage} ,
